+++ backend/providers/tests/test_vidking.py
@@ -2,22 +2,22 @@ from django.test import TestCase
 from providers.resolver import resolve_playback_source, derive_embed_url
 
 
-class VidkingResolverTests(TestCase):
+class EmbedApiResolverTests(TestCase):
     def test_movie_url(self):
         source = resolve_playback_source(
-            provider="vidking",
+            provider="embed-api",
             media_type="movie",
             external_id="1078605",
         )
         url = derive_embed_url(source)
         self.assertEqual(
             url,
-            "https://www.vidking.net/embed/movie/1078605",
+            "https://watch.embed-api.stream/embed/movie/1078605",
         )
 
     def test_tv_url(self):
         source = resolve_playback_source(
-            provider="vidking",
+            provider="embed-api",
             media_type="tv",
             external_id="119051",
             season=1,
@@ -26,5 +26,5 @@ class VidkingResolverTests(TestCase):
         url = derive_embed_url(source)
         self.assertEqual(
             url,
-            "https://www.vidking.net/embed/tv/119051/1/8",
+            "https://watch.embed-api.stream/embed/tv/119051/1/8",
         )